+++ index.html
@@ -58,7 +58,7 @@
             </form>
         </div>
     </main>
-    <script src="js/validation.js" type="module"></script>
-    <script src="js/login.js" type="module"></script>
+    <!-- <script src="js/registration.mjs" type="module"></script> -->
+    <script src="js/login.mjs" type="module"></script>
 </body>
 </html>
