+++ js/imports/url.mjs
@@ -1,0 +1,13 @@
+// creating the API url
+const base = "https://api.noroff.dev/api/v1";
+const socialEndpoints = "/social";
+const authentication = "/auth";
+const register = "/register";
+const login = "/login";
+const profile = "/profiles";
+const post = "/posts";
+
+export const urlLogin = base + socialEndpoints + authentication + login;
+export const urlRegister = base + socialEndpoints + authentication + register;
+export const urlProfile = base + socialEndpoints + profile;
+export const urlPost = base + socialEndpoints + post;
